refactored EventCreate into its own component
